**ChangeLog**: Remove the database creation from the liquibase database schema creation scripts as the basic requirement for a liquibase is to have a database in place.
Updating the `splitStatements` property of each of the sql file to be false.

**Testing**:
Case 1: Ran the `./gradlew update` command consecutively and verified that liquibase skips table creation if the corresponding tables are not present.

Case 2: Deleted all the elr* tables and ran the liquibase update - `./gradlew update`. Verified that the tables are created.