CNDIT-757 : Update the DI Data Definition Language Scripts  (#90)
* Update the DI Data Definition Language Scripts to include the Database creation and selectively create tables only if they are absent.

* Remove the `COLLATE SQL_Latin1_General_CP1_CI_AS` from the column creation in table DDL creation scripts.

* **ChangeLog**:
Remove the database creation from the liquibase database schema creation scripts as the basic requirement for a liquibase is to have a database in place.

Updating the `splitStatements` property of each of the sql file to be false.

**Testing**:
Case 1: Ran the `./gradlew update` command consecutively and verified that liquibase skips table creation if the corresponding tables are not present.

Case 2: Deleted all the elr* tables and ran the liquibase update - `./gradlew update`. Verified that the tables are created.